Add Sprite object
This change adds basic sprites objects which can be moved around the
screen.  Also added was the concept of an "alpha channel" which allows
a rune to be set which is never rendered.  This makes it easy to draw
a rasterized text image with transparent parts so that you can see
the objects underneath.